Extrat map/group view classes into public app for easier sharing
